qcacld-3.0: Update meas_duration type
Currently for sta stats and chan load meas_duration type
is uint8_t. So when duration greater than 254 TU is provided
only first 8 bits get considered as measurement duartion while
populating duration in measureement report.

Fix is to change meas duration type to uint16_t, so that duration
upto 2 octects can be handled.

CRs-Fixed: 3663421

+++ core/mac/inc/sir_mac_prot_def.h
@@ -1391,7 +1391,7 @@ struct chan_load_report {
 	uint8_t op_class;
 	uint8_t channel;
 	qdf_time_t rrm_scan_tsf;
-	uint8_t meas_duration;
+	uint16_t meas_duration;
 	uint8_t chan_load;
 	struct sir_mac_bw_ind_element bw_ind;
 	struct sir_mac_wide_bw_chan_switch wide_bw;
@@ -1488,7 +1488,7 @@ union stats_group_data {
  * @group stats: stats data
  */
 struct statistics_report {
-	uint8_t meas_duration;
+	uint16_t meas_duration;
 	uint8_t group_id;
 	union stats_group_data group_stats;
 };
